Remove extra legacy ops, add new constructors to match code gen

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/shape/OneHot.java
@@ -39,8 +39,8 @@ import java.util.*;
 public class OneHot extends DynamicCustomOp {
     public static final DataType DEFAULT_DTYPE = DataType.FLOAT;
 
-    private int depth;
-    private int jaxis = -1;
+    private long depth;
+    private long jaxis = -1;
     private double on;
     private double off;
     private DataType outputType;
